Fix GeosetType for ChrCustomizationGeosetEntry in DB2Structure

+++ src/server/game/DataStores/DB2Structure.h
@@ -816,7 +816,7 @@ struct ChrCustomizationMaterialEntry
 struct ChrCustomizationGeosetEntry
 {
     uint32 ID;
-    int32 GeosetType;
+    int8 GeosetType;
     int32 GeosetID;
     int32 Modifier;
 };
